restore now saves current session
